fs: quota: restore i_flags of quota files in dquot_disable
When we load quota files, we set the i_flags of the inodes
to (S_NOQUOTA | S_NOATIME | S_IMMUTABLE). In dquot_disable()
we only clear these flags, But we do not restore the old
flags. That would change the i_flags of quota files if some
of these flags were set before loading quota files.

Signed-off-by: Dongsheng Yang <yangds.fnst@cn.fujitsu.com>