rbd: refresh features and set the disk to readonly if there is unsupported bit
If there is unsupported feature enabled after rbd map, we will noticed it in
refreshing, then we need to set the disk readonly.

For example, if we enabled the object-map feature after rbd map, as we don't
support object-map in kernel rbd currently,  we should avoid user to write
rbd device. And set dist to read-write when we disable the unsupported features.

Signed-off-by: Dongsheng Yang <dongsheng.yang@easystack.cn>