ceph: show all options in client_options even if option is equal with default value
As the /sys/kernel/debug/ceph/XXX/client_options only print options
which is different with default values, user are not able to know
the default values of them.

This patch make the /sys/kernel/debug/ceph/XXX/client_options to show
all options even if the option is equal with default value.

Signed-off-by: Dongsheng Yang <dongsheng.yang@easystack.cn>